fix(booking): adjust status verification to reopen room calendar upon booking cancellation

+++ src/main/java/com/stayhub/backend/Module/Property/DTO/Response/SubscriptionPlanResponse.java
@@ -13,7 +13,7 @@ public record SubscriptionPlanResponse(
         BigDecimal price,
         Integer durationMonths,
         Integer maxListings,
-        Double commissionRate,
+        BigDecimal commissionRate,
         BigDecimal creditLimit
 ) {
 }
